scheduler_service/requirements.txt
flask
flask_sqlalchemy
pymysql
python-dotenv
requests